InitArtistList대신 GetArtistList함수로 페이지 및 사이즈로 데이터 가져오기 MyViewModel에서 init으로 실행하는 것이 아닌 HomeScreen에서 실행할 것. 첫 실행은 LaunchedEffect에서 실행
